+++ inc/woocommerce/class-storefront-woocommerce.php
@@ -159,7 +159,7 @@ if ( ! class_exists( 'Storefront_WooCommerce' ) ) :
 			 * Bookings
 			 */
 			if ( $this->is_woocommerce_extension_activated( 'WC_Bookings' ) ) {
-				wp_enqueue_style( 'storefront-woocommerce-bookings-style', get_template_directory_uri() . '/assets/sass/woocommerce/bookings.css', 'storefront-woocommerce-style' );
+				wp_enqueue_style( 'storefront-woocommerce-bookings-style', get_template_directory_uri() . '/assets/sass/woocommerce/extensions/bookings.css', 'storefront-woocommerce-style' );
 				wp_style_add_data( 'storefront-woocommerce-bookings-style', 'rtl', 'replace' );
 			}
 
@@ -167,7 +167,7 @@ if ( ! class_exists( 'Storefront_WooCommerce' ) ) :
 			 * Brands
 			 */
 			if ( $this->is_woocommerce_extension_activated( 'WC_Brands' ) ) {
-				wp_enqueue_style( 'storefront-woocommerce-brands-style', get_template_directory_uri() . '/assets/sass/woocommerce/brands.css', 'storefront-woocommerce-style' );
+				wp_enqueue_style( 'storefront-woocommerce-brands-style', get_template_directory_uri() . '/assets/sass/woocommerce/extensions/brands.css', 'storefront-woocommerce-style' );
 				wp_style_add_data( 'storefront-woocommerce-brands-style', 'rtl', 'replace' );
 			}
 
@@ -175,7 +175,7 @@ if ( ! class_exists( 'Storefront_WooCommerce' ) ) :
 			 * Wishlists
 			 */
 			if ( $this->is_woocommerce_extension_activated( 'WC_Wishlists_Wishlist' ) ) {
-				wp_enqueue_style( 'storefront-woocommerce-wishlists-style', get_template_directory_uri() . '/assets/sass/woocommerce/wishlists.css', 'storefront-woocommerce-style' );
+				wp_enqueue_style( 'storefront-woocommerce-wishlists-style', get_template_directory_uri() . '/assets/sass/woocommerce/extensions/wishlists.css', 'storefront-woocommerce-style' );
 				wp_style_add_data( 'storefront-woocommerce-wishlists-style', 'rtl', 'replace' );
 			}
 
@@ -183,7 +183,7 @@ if ( ! class_exists( 'Storefront_WooCommerce' ) ) :
 			 * AJAX Layered Nav
 			 */
 			if ( $this->is_woocommerce_extension_activated( 'SOD_Widget_Ajax_Layered_Nav' ) ) {
-				wp_enqueue_style( 'storefront-woocommerce-ajax-layered-nav-style', get_template_directory_uri() . '/assets/sass/woocommerce/ajax-layered-nav.css', 'storefront-woocommerce-style' );
+				wp_enqueue_style( 'storefront-woocommerce-ajax-layered-nav-style', get_template_directory_uri() . '/assets/sass/woocommerce/extensions/ajax-layered-nav.css', 'storefront-woocommerce-style' );
 				wp_style_add_data( 'storefront-woocommerce-ajax-layered-nav-style', 'rtl', 'replace' );
 			}
 
@@ -191,7 +191,7 @@ if ( ! class_exists( 'Storefront_WooCommerce' ) ) :
 			 * Variation Swatches
 			 */
 			if ( $this->is_woocommerce_extension_activated( 'WC_SwatchesPlugin' ) ) {
-				wp_enqueue_style( 'storefront-woocommerce-variation-swatches-style', get_template_directory_uri() . '/assets/sass/woocommerce/variation-swatches.css', 'storefront-woocommerce-style' );
+				wp_enqueue_style( 'storefront-woocommerce-variation-swatches-style', get_template_directory_uri() . '/assets/sass/woocommerce/extensions/variation-swatches.css', 'storefront-woocommerce-style' );
 				wp_style_add_data( 'storefront-woocommerce-variation-swatches-style', 'rtl', 'replace' );
 			}
 
@@ -199,7 +199,7 @@ if ( ! class_exists( 'Storefront_WooCommerce' ) ) :
 			 * Composite Products
 			 */
 			if ( $this->is_woocommerce_extension_activated( 'WC_Composite_Products' ) ) {
-				wp_enqueue_style( 'storefront-woocommerce-composite-products-style', get_template_directory_uri() . '/assets/sass/woocommerce/composite-products.css', 'storefront-woocommerce-style' );
+				wp_enqueue_style( 'storefront-woocommerce-composite-products-style', get_template_directory_uri() . '/assets/sass/woocommerce/extensions/composite-products.css', 'storefront-woocommerce-style' );
 				wp_style_add_data( 'storefront-woocommerce-composite-products-style', 'rtl', 'replace' );
 			}
 
@@ -207,7 +207,7 @@ if ( ! class_exists( 'Storefront_WooCommerce' ) ) :
 			 * WooCommerce Photography
 			 */
 			if ( $this->is_woocommerce_extension_activated( 'WC_Photography' ) ) {
-				wp_enqueue_style( 'storefront-woocommerce-photography-style', get_template_directory_uri() . '/assets/sass/woocommerce/photography.css', 'storefront-woocommerce-style' );
+				wp_enqueue_style( 'storefront-woocommerce-photography-style', get_template_directory_uri() . '/assets/sass/woocommerce/extensions/photography.css', 'storefront-woocommerce-style' );
 				wp_style_add_data( 'storefront-woocommerce-photography-style', 'rtl', 'replace' );
 			}
 
@@ -215,7 +215,7 @@ if ( ! class_exists( 'Storefront_WooCommerce' ) ) :
 			 * Product Reviews Pro
 			 */
 			if ( $this->is_woocommerce_extension_activated( 'WC_Product_Reviews_Pro' ) ) {
-				wp_enqueue_style( 'storefront-woocommerce-product-reviews-pro-style', get_template_directory_uri() . '/assets/sass/woocommerce/product-reviews-pro.css', 'storefront-woocommerce-style' );
+				wp_enqueue_style( 'storefront-woocommerce-product-reviews-pro-style', get_template_directory_uri() . '/assets/sass/woocommerce/extensions/product-reviews-pro.css', 'storefront-woocommerce-style' );
 				wp_style_add_data( 'storefront-woocommerce-product-reviews-pro-style', 'rtl', 'replace' );
 			}
 
@@ -223,7 +223,7 @@ if ( ! class_exists( 'Storefront_WooCommerce' ) ) :
 			 * WooCommerce Smart Coupons
 			 */
 			if ( $this->is_woocommerce_extension_activated( 'WC_Smart_Coupons' ) ) {
-				wp_enqueue_style( 'storefront-woocommerce-smart-coupons-style', get_template_directory_uri() . '/assets/sass/woocommerce/smart-coupons.css', 'storefront-woocommerce-style' );
+				wp_enqueue_style( 'storefront-woocommerce-smart-coupons-style', get_template_directory_uri() . '/assets/sass/woocommerce/extensions/smart-coupons.css', 'storefront-woocommerce-style' );
 				wp_style_add_data( 'storefront-woocommerce-smart-coupons-style', 'rtl', 'replace' );
 			}
 
@@ -231,7 +231,7 @@ if ( ! class_exists( 'Storefront_WooCommerce' ) ) :
 			 * WooCommerce Deposits
 			 */
 			if ( $this->is_woocommerce_extension_activated( 'WC_Deposits' ) ) {
-				wp_enqueue_style( 'storefront-woocommerce-deposits-style', get_template_directory_uri() . '/assets/sass/woocommerce/deposits.css', 'storefront-woocommerce-style' );
+				wp_enqueue_style( 'storefront-woocommerce-deposits-style', get_template_directory_uri() . '/assets/sass/woocommerce/extensions/deposits.css', 'storefront-woocommerce-style' );
 				wp_style_add_data( 'storefront-woocommerce-deposits-style', 'rtl', 'replace' );
 			}
 
@@ -239,7 +239,7 @@ if ( ! class_exists( 'Storefront_WooCommerce' ) ) :
 			 * WooCommerce Product Bundles
 			 */
 			if ( $this->is_woocommerce_extension_activated( 'WC_Bundles' ) ) {
-				wp_enqueue_style( 'storefront-woocommerce-bundles-style', get_template_directory_uri() . '/assets/sass/woocommerce/bundles.css', 'storefront-woocommerce-style' );
+				wp_enqueue_style( 'storefront-woocommerce-bundles-style', get_template_directory_uri() . '/assets/sass/woocommerce/extensions/bundles.css', 'storefront-woocommerce-style' );
 				wp_style_add_data( 'storefront-woocommerce-bundles-style', 'rtl', 'replace' );
 			}
 
@@ -247,7 +247,7 @@ if ( ! class_exists( 'Storefront_WooCommerce' ) ) :
 			 * WooCommerce Multiple Shipping Addresses
 			 */
 			if ( $this->is_woocommerce_extension_activated( 'WC_Ship_Multiple' ) ) {
-				wp_enqueue_style( 'storefront-woocommerce-sma-style', get_template_directory_uri() . '/assets/sass/woocommerce/ship-multiple-addresses.css', 'storefront-woocommerce-style' );
+				wp_enqueue_style( 'storefront-woocommerce-sma-style', get_template_directory_uri() . '/assets/sass/woocommerce/extensions/ship-multiple-addresses.css', 'storefront-woocommerce-style' );
 				wp_style_add_data( 'storefront-woocommerce-sma-style', 'rtl', 'replace' );
 			}