+++ changelog.md
@@ -3,6 +3,7 @@
 ### *2016.xx.xx* - 2.1.0
 * **New** - Lots of SEO enhancements. Kudos [@opportus](https://github.com/opportus).
 * **New** - Selective refresh on site title / tag line.
+* **Fix** - Product thumbnail size on cart page.
 
 ### *2016.07.26* - 2.0.7
 * **Fix** - Compatibility with PHP 5.2.